requirements.txt
pandas
numpy
yfinance
pandas_ta
xgboost
scikit-learn
plotly
python-dotenv
ta
python-telegram-bot
discord.py
mercadopago
sqlalchemy
requests
